Restructures the project files. Creates 'utils' and 'file_load' modules to store all related components.
