+++ file_load/meta_file_text.py
@@ -1,163 +1,15 @@
 # import pyodbc
 import sys
-import os
-import time
-import traceback
 from pathlib import Path
-import xlrd #installation: pip install xlrd
 import json
 from collections import OrderedDict
-import file_errors as ferr # custom library containing all error processing related classes
-import db_access as db # custom library containing all database related classes
-import logging
-import mdl_logging as ml
-import main_cfg as mc
-import global_const as gc
-
-# common_logger_name = 'file_processing_log'
-# logging_level = 'INFO'
-
-class StudyConfig:
-	config_loc = None
-	config_glb = None
-	study_logger_name =  ''# 'file_processing_log'
-	study_logging_level = '' #'INFO'
-
-
-class FieldIdMethod:
-	field_id_methods = ['name', 'number']
-	name = field_id_methods[0]
-	number = field_id_methods[1]
-
-
-#Text file class (used as a base)
-class File:
-	filepath = None
-	wrkdir = None
-	filename = None
-	file_type = None #1:text, 2:excel
-	file_delim = None # ','
-	lineList = None # []
-	__headers = None # []
-	error = None  # FileErrors class reference holding all errors associated with the current file
-	sample_id_field_names = None # []
-	loaded = None
-	logger = None
-
-	def __init__(self, filepath, file_type = 1, file_delim = ','):
-		self.filepath = filepath
-		self.wrkdir = os.path.dirname(os.path.abspath(filepath))
-		self.filename = Path(os.path.abspath(filepath)).name
-		self.file_type = file_type
-		self.file_delim = file_delim
-		self.error = ferr.FileError(self)
-		self.lineList = []
-		self.__headers = []
-		self.sample_id_field_names = []
-		self.loaded = False
-
-	@property
-	def headers(self):
-		if not self.__headers:
-			self.getHeaders()
-		return self.__headers
-
-	def setup_logger(self, wrkdir, filename):
-
-		lg = ml.setup_logger(StudyConfig.study_logger_name, StudyConfig.study_logging_level,
-					Path(wrkdir) / 'Logs',
-					filename + '_' + time.strftime("%Y%m%d_%H%M%S", time.localtime()) + '.log')
-
-		self.log_handler = lg['handler']
-		return lg['logger']
-
-	def getFileContent (self):
-		if not self.logger:
-			loc_log = logging.getLogger(StudyConfig.study_logger_name)
-		else:
-			loc_log = self.logger
-
-		if not self.lineList:
-			if self.fileExists (self.filepath):
-				loc_log.debug ('Loading file content of "{}"'.format(self.filepath))
-				fl = open(self.filepath, "r")
-				self.lineList = [line.rstrip('\n') for line in fl]
-				fl.close()
-				self.loaded = True
-			else:
-				_str = 'Loading content of the file "{}" failed since the file does not appear to exist".'.format(self.filepath)
-				self.error.addError(_str)
-				loc_log.error(_str)
-				self.lineList = None
-				self.loaded = False
-		return self.lineList
-
-	def fileExists(self, fn):
-		try:
-			open(fn, "r")
-			return 1
-		except IOError:
-			return 0
-
-	def getHeaders (self):
-		if not self.__headers:
-			hdrs = self.getRowByNumber(1).split(self.file_delim)
-			self.__headers = [hdr.strip().replace(' ', '_') for hdr in hdrs]
-		return self.__headers
-
-	def getRowByNumber (self, rownum):
-		lineList = self.getFileContent()
-		#check that requested row is withing available records of the file and >0
-		if not lineList == None and len(lineList) >= rownum and rownum > 0:
-			return lineList[rownum-1]
-		else:
-			return ""
-
-	def rowsCount(self, excludeHeader = False):
-		num = len(self.getFileContent())
-		if excludeHeader:
-			num = num - 1
-		return num
-
-#Config file class
-class ConfigFile(File):
-	config_items = None
-	# config_items_populated = False
-	key_value_delim = None
-	line_comment_sign = None
-
-	def __init__(self, filepath, file_type=1, key_value_delim=':', line_comment_sign='##'):
-		self.key_value_delim = key_value_delim
-		self.line_comment_sign = line_comment_sign
-		self.config_items = {}
-		File.__init__(self, filepath, file_type, self.key_value_delim)
-		self.loadConfigSettings()
-
-	# loads config setting assuming that it is a dictionary ==> key: value
-	# to identify all key/value pairs, it will split based on ":" for 1 delimiter only and will keep the rest as value of the key
-	# any text after "##" will be considered a comment and will be ignored
-	def loadConfigSettings(self):
-		if not self.config_items: # self.config_items_populated:
-			lns = File.getFileContent(self)
-
-			if self.loaded:
-				for l in lns:
-					values = l.split(self.line_comment_sign, 1)[0].split(self.file_delim, 1)  # use only first delimiter
-
-					if len(values) >= 2:
-						# add items to a dictionary
-						self.config_items[values[0].strip()] = values[1].strip()
-		return self.config_items
-
-	def getItemByKey(self, item_key):
-		# check if config file was already loaded
-		if not self.config_items:
-			self.loadConfigSettings()
-		# check if requested key exists
-		if item_key in self.config_items:
-			return self.config_items[item_key]
-		else:
-			return None
+from file_load.file_errors import RowErrors
+from utils import MetadataDB # database connectivity related class
+from utils.mdl_logging import deactivate_logger_common
+from utils import global_const as gc
+from .file import File
+from .file_utils import StudyConfig, FieldIdMethod, loadConfiguration
+from .row import Row
 
 # metadata text file class
 class MetaFileText(File):
@@ -281,7 +133,7 @@ class MetaFileText(File):
 		# print('file name through Error object - getFileRow() - before Row instance = {}'.format(self.error.entity.filepath))
 
 		row = Row(self, rownum, lst_content, hdrs)
-		row.error = ferr.RowErrors(row)
+		row.error = RowErrors(row)
 
 		if len(hdrs) == len (lst_content):
 			self._validateMandatoryFieldsPerRow(row) # validate row for required fields being populated
@@ -478,7 +330,7 @@ class MetaFileText(File):
 					# print ('No Errors - Saving to DB, Row Info: {}'.format (row.toStr()))
 					self.logger.info('Row #{}. No Row level errors were identified. Saving it to database. Row data: {}'.format(
 						row.row_number, row.toStr()))
-					mdb = db.MetadataDB(self.cfg_file)
+					mdb = MetadataDB(self.cfg_file)
 
 					mdb_resp = mdb.submitRow(row, self) # row.sample_id, row.toJSON(), self.getFileDictionary_JSON(True), str(self.filepath))
 					if not row.error.errorsExist():
@@ -517,270 +369,10 @@ class MetaFileText(File):
 
 		# release the log handler for the current file
 		#self.logger.removeHandler(self.log_handler)
-		ml.deactivate_logger(self.logger, self.log_handler)
+		deactivate_logger_common(self.logger, self.log_handler)
 
 # TODO: for Text and Excel files - handling commas a part of the field values provided.
 #  		Idea is to accomodate double quotes as text identifier; however double quotes should not be considered a value
-
-# metadata Excel file class
-class MetaFileExcel(MetaFileText):
-	# cfg_file = None
-	# file_dict = None  # OrderedDict()
-	# rows = None  # OrderedDict()
-	sheet_name = None
-
-	def __init__(self, filepath, cfg_path = '', file_type=2, sheet_name = ''):
-
-		#loadConfiguration (main_cfg_obj) # load global and local configureations
-
-		File.__init__(self, filepath, file_type)
-
-		self.logger = self.setup_logger(self.wrkdir, self.filename)
-		self.logger.info('Start working with file {}'.format(filepath))
-
-		self.logger.info('Loading config file.')
-		# identify name of the config file for a study
-		if len(cfg_path) == 0:
-			cfg_path = Path(self.wrkdir) / gc.default_study_config_file
-
-		if self.textFileExists(cfg_path):
-			loadConfiguration(self, cfg_path)
-			self.cfg_file = StudyConfig.config_loc
-			self.file_dict = OrderedDict()
-			self.rows = OrderedDict()
-
-			self.sheet_name = ''
-			# self.sheet_name = sheet_name.strip()
-			if not self.sheet_name or len(self.sheet_name) == 0:
-				# if sheet name was not passed as a parameter, try to get it from config file
-				self.sheet_name = self.cfg_file.getItemByKey(gc.study_excel_wk_sheet_name) # 'wk_sheet_name'
-			# print (self.sheet_name)
-			self.logger.info('Sheet name that data will be loaded from: "{}"'.format(self.sheet_name))
-		else:
-			_str = 'Study configuration file "{}" does not exist, configuration loading was aborted.'.format(cfg_path)
-			self.error.addError(_str)
-			self.logger.error(_str)
-
-		'''	
-		# self.cfg_file = None
-		cfg_file = self.getConfigInfo ('') # '(cfg_filepath)
-		if not cfg_file.loaded:
-			# report error for for failed config file loading
-			_str = 'Neither the provided config file "{}" nor default "config.cfg" file could not be loaded.'.format(cfg_filepath)
-			self.error.addError(_str)
-			self.logger.error(_str)
-
-		self.file_dict = OrderedDict()
-		self.rows = OrderedDict()
-		self.sheet_name = ''
-		# self.sheet_name = sheet_name.strip()
-		if not self.sheet_name or len(self.sheet_name) == 0:
-			# if sheet name was not passed as a parameter, try to get it from config file
-			self.sheet_name = self.cfg_file.getItemByKey('wk_sheet_name')
-			# print (self.sheet_name)
-		self.logger.info('Sheet name that data will be loaded from: "{}"'.format(self.sheet_name))
-		'''
-
-	def getFileContent (self):
-		if not self.lineList:
-			if self.fileExists (self.filepath):
-				self.logger.debug('Loading file content of "{}"'.format(self.filepath))
-
-				wb = xlrd.open_workbook(self.filepath)
-				if not self.sheet_name or len(self.sheet_name) == 0:
-					# by default retrieve the first sheet in the excel file
-					sheet = wb.sheet_by_index(0)
-				else:
-					# if sheet name was provided
-					sheets = wb.sheet_names() # get list of all sheets
-					if (self.sheet_name in sheets):
-						# if given sheet name in the list of available sheets, load the sheet
-						sheet = wb.sheet_by_name(self.sheet_name)
-					else:
-						# report an error if given sheet name not in the list of available sheets
-						_str = 'Given sheet name "{}" was not found in the file "{}". Verify that the sheet name exists in the file.'.format(self.sheet_name, self.filepath)
-						self.error.addError(_str)
-						self.logger.error(_str)
-
-						self.lineList = None
-						self.loaded = False
-						return self.lineList
-
-				sheet.cell_value(0, 0)
-
-				for i in range(sheet.nrows):
-					# ln = sheet.row_values(i)
-					# print (ln)
-
-					ln = []
-					for j in range(sheet.ncols):
-						# print(sheet.cell_value(i, j))
-						# ln.append('"' + sheet.cell_value(i,j) + '"')
-						cell = sheet.cell(i, j)
-						cell_value = cell.value
-						# take care of number and dates received from Excel and converted to float by default
-						if cell.ctype == 2 and int(cell_value) == cell_value:
-							# the value is integer
-							cell_value = str(int(cell_value))
-						elif cell.ctype == 2:
-							# the value is float
-							cell_value = str(cell_value)
-						# convert date back to human readable date format
-						# print ('cell_value = {}'.format(cell_value))
-						if cell.ctype == 3:
-							cell_value_date = xlrd.xldate_as_datetime(cell_value, wb.datemode)
-							cell_value = cell_value_date.strftime("%Y-%m-%d")
-						ln.append(cell_value)
-
-					self.lineList.append (','.join(ln))
-
-				wb.unload_sheet(sheet.name)
-				self.loaded = True
-			else:
-				_str = 'Loading content of the file "{}" failed since the file does not appear to exist".'.format(self.filepath)
-				self.error.addError(_str)
-				self.logger.error(_str)
-
-				self.lineList = None
-				self.loaded = False
-		return self.lineList
-
-	def textFileExists(self, fn):
-		return MetaFileText.fileExists(self, fn)
-
-	def fileExists(self, fn):
-		try:
-			wb = xlrd.open_workbook(fn)
-			return 1
-		except Exception as ex: # IOError
-			# print (ex)
-			return 0
-
-class Row ():
-	file = None #reference to the file object that this row belongs to
-	row_number = None #row number - header = row #1, next line #2, etc.
-	row_content = None #[] #list of values from a file for this row
-	_row_dict = None #OrderedDict() of a headers
-	header = None # [] #list of values from a file for the first row (headers)
-	__error = None #RowErrors class reference holding all errors associated with the current row
-	__sample_id = None #it stores a sample Id value for the row.
-
-	def __init__(self, file, row_num, row_content, header):
-		self.file = file
-		self.row_number = row_num
-		self.row_content = row_content
-		self.header = header
-		self.row_dict = OrderedDict()
-
-	@property
-	def sample_id(self):
-		return self.__sample_id
-
-	@sample_id.setter
-	def sample_id(self, value):
-		self.__sample_id = value
-
-	@property
-	def row_dict (self):
-		return self._row_dict
-
-	@row_dict.setter
-	def row_dict(self, value):
-		self._row_dict = value
-
-	@property
-	def error(self):
-		return self.__error
-
-	@error.setter
-	def error(self, value):
-		self.__error = value
-
-	def toJSON(self):
-		# print ('From withing toJSON - Dictionary source:{}'.format(self.row_dict))
-		return json.dumps(self.row_dict)
-
-	def toStr(self):
-		row = {
-			'file':str(self.file.filepath),
-			'row_number':self.row_number,
-			'sample_id':self.sample_id,
-			'row_JSON':self.toJSON(),
-			'errors': self.error.errors
-		}
-		return row
-
-	def assignSampleId(self):
-		self.file.logger.debug('Row #{}. Assigning sample id value.'.format(self.row_number))
-
-		cfg = self.file.getConfigInfo()
-		delim = self.file.configValueListSeparator()
-
-		# retrieve config values for sample id retrieval
-		sid = cfg.getItemByKey('sample_id_expression').strip()
-		fields = cfg.getItemByKey('sample_id_fields').split(delim)
-		method = cfg.getItemByKey('sample_id_method').strip() #split(delim)[0].
-
-		for sf in fields:
-			i = 0  # keeps field count
-			for hdr, cnt in zip(self.header, self.row_content):
-				i += 1
-				# check sample_id fields
-				if method == FieldIdMethod.name:  # self.field_id_methods[0]: # 'name':
-					smp_val = hdr.strip()
-				elif method == FieldIdMethod.number: # self.field_id_methods[1]: # 'number':
-					smp_val = i
-				else:
-					smp_val = None
-
-				if str(smp_val) == sf.strip():
-					sid = sid.replace('{{{}}}'.format(str(smp_val)), '"' + cnt + '"')
-
-		self.file.logger.debug ('Row #{}. Expression for sample id evaluation: "{}"'.format(self.row_number, sid))
-		try:
-			smp_evaled = eval(sid) # attempt to evaluate expression for sample id
-		except Exception as ex:
-			# report an error if evaluation has failed.
-			_str = 'Error "{}" occurred during evaluating sample id expression: {}\n{} '.format(ex, sid, traceback.format_exc())
-			self.error.addError(_str)
-			self.file.logger.error(_str)
-			self.file.logger.debug(sys.exc_info()[1])
-			smp_evaled = ''
-
-		# TODO: add validation for duplicated Sample IDs. If duplication is found, reject the row.
-
-		self.__sample_id = str(smp_evaled).strip()
-		return self.__sample_id
-
-def loadConfiguration (fl_class, loc_cfg_path):
-	# load global configuration
-
-	#loc_log = logging.getLogger(StudyConfig.study_logger_name)
-	m_cfg = mc.ConfigData(gc.main_config_file)
-	m_logger_name = m_cfg.get_value('Logging/main_log_name')
-	m_logger = logging.getLogger(m_logger_name)
-
-	m_logger.debug('Loading Global config file {} for file: {}'.format(gc.main_config_file, fl_class.filepath))
-	StudyConfig.config_glb = mc.ConfigData(gc.main_config_file)
-
-	m_logger.info('Loading Study config file {} for file: {}'.format(loc_cfg_path, fl_class.filepath))
-	# load local configuration
-	try:
-		StudyConfig.config_loc = mc.ConfigData(loc_cfg_path)
-	except Exception as ex:
-		m_logger.error('Error "{}" occurred during loading study config file "{}"\n{}'.format(
-			ex, loc_cfg_path, traceback.format_exc()))
-		#raise
-		return False
-
-	#load global logging setting
-	StudyConfig.study_logger_name = StudyConfig.config_glb.get_value(gc.study_logger_name_cfg_path)
-	StudyConfig.study_logging_level = StudyConfig.config_glb.get_value(gc.study_logging_level_cfg_path)
-
-	return True
-
-
-	# populate values from local config file
 
 
 # if executed by itself, do the following
