+++ utils/configuration.py
@@ -3,7 +3,6 @@ import yaml
 class ConfigData:
 
     def __init__(self, cfg_path):
-    # def load_yaml_config(cfg_path):
         self.loaded = False
         with open(cfg_path, 'r') as ymlfile:
             self.cfg = yaml.load(ymlfile)
@@ -22,9 +21,6 @@ class ConfigData:
     def getItemByKey (self, key_name):
         return str(self.get_value(key_name))
 
-    #def get_cfg_value (cfg_ref, section, key):
-        #return cfg_ref[section][key]
-
 '''    
 for section in cfg:
     print(section)
@@ -34,6 +30,5 @@ for section in cfg:
 #print(cfg['DB'])
 #print(cfg['Logging'])
 '''
-
-#ym = load_yaml_config('E:\MounSinai\MoTrPac_API\ProgrammaticConnectivity\MountSinai_metadata_file_loader\DataFiles\study01\study.cfg_1.yaml')
-#print (ym)
+# ym = load_yaml_config('E:\MounSinai\MoTrPac_API\ProgrammaticConnectivity\MountSinai_metadata_file_loader\DataFiles\study01\study.cfg_1.yaml')
+# print (ym)
