+++ utils/global_const.py
